fix: Fixed all the components file
