+++ src/components/Card/index.ts
@@ -1,4 +1,4 @@
 export { default as Card } from './Card';
-export { default as ArrowButton } from './ArrowButton';
-export { default as CardInformation } from './CardInformation';
+export { default as ArrowButton } from './CardButton/ArrowButton';
+export { default as CardInformation } from './CardInformation/CardInformation';
 export { CustomSlider } from './CardUtils';
