+++ src/components/Detail/DetailButton/DetailButtonUtils.ts
@@ -1,0 +1,28 @@
+import styled from 'styled-components';
+
+export const DetailButtonContainer = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 30%;
+`;
+
+export const DetailPickButtonContainer = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border-radius: 10px;
+  padding: 0.5rem 0.1rem;
+  font-size: 1.2rem;
+  background-color: #f2f2f2;
+`;
+
+export const DetailPickButton = styled.img`
+  width: 30%;
+  margin-right: 0.5rem;
+`;
+
+export const DetailInformationTooltip = styled.img`
+  width: 20%;
+  margin-left: 0.5rem;
+`;
