+++ src/components/Detail/DetailRate/DetailRateUtils.ts
@@ -1,0 +1,20 @@
+import styled from 'styled-components';
+
+export const DetailRateContainer = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: flex-start;
+  width: 100%;
+  height: 5%;
+  font-size: 1.5rem;
+`;
+
+export const DetailRateStar = styled.div`
+  width: 40%;
+  height: 100%;
+  background-color: #f2f2f2;
+`;
+
+export const DetailRateNumsReview = styled.div`
+  font-size: 0.8rem;
+`;
