+++ src/components/Detail/index.ts
@@ -3,19 +3,10 @@ export {
   DetailDrawer,
   BackButtonContainer,
   DetailTopContainer,
-  DetailRateContainer,
-  DetailImageContainer,
   DetailGPTContainer,
   DetailBackButton,
   DetailTitle,
-  DetailButtonContainer,
-  DetailPickButtonContainer,
-  DetailPickButton,
-  DetailInformationTooltip,
-  DetailRateStar,
-  DetailRateNumsReview,
-  DetailImg,
 } from './DetailUtil';
-export { default as DetailButton } from './DetailButton';
-export { default as DetailRate } from './DetailRate';
-export { default as DetailImage } from './DetailImage';
+export { default as DetailButton } from './DetailButton/DetailButton';
+export { default as DetailRate } from './DetailRate/DetailRate';
+export { default as DetailImage } from './DetailImage/DetailImage';
