+++ src/components/Header/SearchBar/SearchBarUtils.ts
@@ -1,6 +1,6 @@
-import styled from 'styled-components';
+import { styled } from 'styled-components';
 
-const SearchBarContainer = styled.div`
+export const SearchBarContainer = styled.div`
   display: flex;
   width: 80%;
   height: 43px;
@@ -11,7 +11,7 @@ const SearchBarContainer = styled.div`
   background-color: ${(props) => props.theme.backgroundTertiary};
 `;
 
-const SearchBarTextContainer = styled.input`
+export const SearchBarTextContainer = styled.input`
   width: 85%;
   font-size: 1.1rem;
   border-style: none;
@@ -19,30 +19,14 @@ const SearchBarTextContainer = styled.input`
   background-color: ${(props) => props.theme.backgroundTertiary};
 `;
 
-const SearchBarButton = styled.button`
+export const SearchBarButton = styled.button`
   display: flex;
   justify-content: center;
   align-items: center;
   width: 15%;
 `;
 
-const SearchIconImg = styled.img`
+export const SearchIconImg = styled.img`
   width: 60%;
   height: 60%;
 `;
-
-const SearchBar = () => {
-  return (
-    <SearchBarContainer>
-      <SearchBarTextContainer placeholder="Search Here" />
-      <SearchBarButton>
-        <SearchIconImg
-          src={`${process.env.PUBLIC_URL}/assets/images/search_icon.svg`}
-          alt="Search Icon"
-        />
-      </SearchBarButton>
-    </SearchBarContainer>
-  );
-};
-
-export default SearchBar;
