+++ src/components/Header/index.ts
@@ -1,4 +1,4 @@
 export { default as Header } from './Header';
-export { default as GoogleFormBanner } from './GoogleFormBanner';
-export { default as SearchBar } from './SearchBar';
-export { default as StaticHashTag } from './StaticHashTag';
+export { default as GoogleFormBanner } from './GoogleFormBanner/GoogleFormBanner';
+export { default as SearchBar } from './SearchBar/SearchBar';
+export { default as StaticHashTag } from './StaticHashTag/StaticHashTag';
