+++ src/components/List/ListUtils.ts
@@ -2,7 +2,8 @@ import styled from 'styled-components';
 
 export const ListContainer = styled.div`
   display: flex;
-  flex-direction: column;
+  flex-wrap: wrap;
+  justify-content: center;
   width: 100%;
   height: 100vh;
   overflow: auto;
