+++ src/components/OnBoarding/OnBoardingBriefContent/OnBoardingBriefContent.tsx
@@ -1,0 +1,11 @@
+import { OnBoardingOneLineContainer, EmphasisText } from './OnBoardingBriefContentUtils';
+
+const OnBoardingBriefContent = () => {
+  return (
+    <OnBoardingOneLineContainer>
+      Find your&nbsp;<EmphasisText>Secret</EmphasisText>&nbsp;cafe
+    </OnBoardingOneLineContainer>
+  );
+};
+
+export default OnBoardingBriefContent;
