+++ src/components/OnBoarding/OnBoardingBriefContent/OnBoardingBriefContentUtils.ts
@@ -1,0 +1,17 @@
+import styled from 'styled-components';
+
+export const OnBoardingOneLineContainer = styled.div`
+  display: flex;
+  align-items: end;
+  text-align: left;
+  width: 100%;
+  height: 20%;
+  margin: 0 0 1rem 0;
+  font-size: 2rem;
+  font-weight: 400;
+`;
+
+export const EmphasisText = styled.div`
+  text-emphasis: filled #ff0000;
+  padding: 0;
+`;
