+++ src/components/OnBoarding/OnBoardingLogo/OnBoardingLogoUtils.ts
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+export const OnBoardingLogoContainer = styled.div`
+  width: 100%;
+  height: 30%;
+`;
+
+export const OnBoardingLogoImg = styled.img`
+  width: 100%;
+  height: 100%;
+`;
