+++ src/components/OnBoarding/index.ts
@@ -1,4 +1,4 @@
 export { default as OnBoarding } from './OnBoarding';
-export { default as OnBoardingBriefContent } from './OnBoardingBriefContent';
-export { default as OnBoardingLogo } from './OnBoardingLogo';
-export { default as OnBoardingFeatures } from './OnBoardingFeatures';
+export { default as OnBoardingBriefContent } from './OnBoardingBriefContent/OnBoardingBriefContent';
+export { default as OnBoardingLogo } from './OnBoardingLogo/OnBoardingLogo';
+export { default as OnBoardingFeatures } from './OnBoardingFeatures/OnBoardingFeatures';
