+++ src/components/Rank/RankUtil.ts
@@ -2,7 +2,7 @@ import styled from 'styled-components';
 
 export const RankingLayouts = styled.div`
   display: flex;
-  flex-direction: column;
+  flex-wrap: wrap;
   width: 100%;
   heigth: 100%;
 `;
@@ -16,10 +16,14 @@ export const RankingHeader = styled.div`
   padding: 0 0 0.5rem 1rem;
 `;
 
+export const RankBackButton = styled.img`
+  width: 8%;
+`;
+
 export const RankingContainer = styled.div`
   display: flex;
-  flex-direction: column;
-  align-items: center;
+  flex-wrap: wrap;
+  justify-content: center;
   width: 100%;
   height: 43%;
 `;
@@ -32,8 +36,4 @@ export const RankingText = styled.div`
   height: 13%;
   padding: 0 3.5rem;
   font-size: 1.4rem;
-`;
-
-export const RankBackButton = styled.img`
-  width: 8%;
 `;
